@@ -7,35 +7,35 @@
 
 namespace App\Logger;
 
+use Monolog\LogRecord;
+
 /**
  * Class TraceIdProcessor.
  */
 class TraceIdProcessor
 {
-    private $traceId;
-
     /**
      * TraceIdProcessor constructor.
      *
-     * @param string $bindTraceId
+     * @param string $traceId
      */
-    public function __construct(string $bindTraceId)
-    {
-        $this->traceId = $bindTraceId;
+    public function __construct(
+        private readonly string $traceId
+    ) {
     }
 
     /**
      * Magic invoke function.
      *
-     * @param array $record
+     * @param LogRecord $record
      *   Log record
      *
-     * @return array
+     * @return LogRecord
      *   The record added require id in extras
      */
-    public function __invoke(array $record)
+    public function __invoke(LogRecord $record): LogRecord
     {
-        $record['extra']['traceId'] = $this->traceId;
+        $record->extra['traceId'] = $this->traceId;
 
         return $record;
     }
